Fix incompatible function argument types

+++ WinQuake/sv_move.cpp
@@ -318,7 +318,7 @@ void SV_NewChaseDir (edict_t *actor, edict_t *enemy, float dist)
 	}
 
 // try other directions
-	if ( ((rand()&3) & 1) ||  abs(deltay)>abs(deltax))
+	if ( ((rand()&3) & 1) ||  abs( static_cast<int>( deltay ) )>abs( static_cast<int>( deltax ) ))
 	{
 		tdir=d[1];
 		d[1]=d[2];
